Fix compile issues with newer versions of Arduino
* New versions of Arduino look like they changed a type and added the
requirement that program memory variables are declared const. The const
declarations are good practice anyway for anything stored in flash
memory.
* Commented out code which generated an error when the serial buffers
were not equal to 128. The requirement for this buffer size is poorly
justified and it seems like it may be a hack to work around a specific
issue which likely could be solved another way. We should try to solve
the issue rather than place requirements on built in Arduino libraries.